Created a filtering file
In this file, i read the data from the Google Drive, then I filter according to the given ranges by the medical experts. GIve a 1 label if they don't meet the criteria and the same for the kidney problems. then exporting this data so the other team members can use it.